requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
scikit-learn
pandas
numpy
openai
anthropic
httpx
streamlit
requests
